Refactor search.json test; assert urls in sitemap.
