+++ package.json
@@ -19,6 +19,7 @@
   },
   "devDependencies": {
     "a11y": "^0.5.1",
+    "fast-xml-parser": "^3.16.0",
     "http-server": "^0.12.1",
     "jest": "^25.1.0",
     "jest-puppeteer": "^4.4.0",
